feat(agent): Debugging agent refactor [wip]
