+++ .vscode/launch.json
@@ -13,7 +13,8 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}\\**\\*.test.ts"
+                "${workspaceFolder}\\**\\demo.test.ts"
+                // "${workspaceFolder}\\**\\*.test.ts"
             ],
             "console": "integratedTerminal",
             "internalConsoleOptions": "neverOpen",
