+++ src/agents/agent-model.ts
@@ -9,7 +9,13 @@ export const isAgent = (o: any): o is Agent =>
 
 export const inactiveAgentProxy = (agent: Agent): Agent =>
     new Proxy(agent, {
+        tempValues: {},
+
         get(target: Agent, property: PropertyKey) {
+            if (property === "tempValues") {
+                return this.tempValues;
+            }
+
             if (property !== "id" && !ContextManager.isContextStatic()) {
                 throw new RegalError(
                     "The properties of an inactive agent cannot be accessed within a game cycle."
@@ -20,16 +26,24 @@ export const inactiveAgentProxy = (agent: Agent): Agent =>
         },
 
         set(target: Agent, property: PropertyKey, value: any) {
-            if (!ContextManager.isContextStatic()) {
-                // TODO - start looking here
-                if (Reflect.get(target, property) !== undefined) {
-                    throw new RegalError(
-                        "The properties of an inactive agent cannot be set within a game cycle."
-                    );
-                }
+            if (ContextManager.isContextStatic()) {
+                return Reflect.set(target, property, value);
+            } else if (StaticAgentRegistry.hasAgent(target.id)) {
+                throw new RegalError(
+                    "This static agent must be activated before it may be modified."
+                );
             }
 
-            return Reflect.set(target, property, value);
+            // Allow initial values to be set (like from a constructor) but ONLY ONCE.
+            if (this.tempValues[property] !== undefined) {
+                throw new RegalError(
+                    "The properties of an inactive agent cannot be set within a game cycle."
+                );
+            }
+
+            this.tempValues[property] = value;
+
+            return true;
         },
 
         deleteProperty(target: Agent, property: PropertyKey) {
@@ -41,12 +55,14 @@ export const inactiveAgentProxy = (agent: Agent): Agent =>
 
             return Reflect.deleteProperty(target, property);
         }
-    });
+    } as ProxyHandler<Agent>);
 
 export const activeAgentProxy = (id: number, game: GameInstance): Agent =>
     new Proxy({} as Agent, {
         get(target: Agent, property: PropertyKey) {
-            return game.agents.getAgentProperty(id, property);
+            return target.hasOwnProperty(property)
+                ? game.agents.getAgentProperty(id, property)
+                : Reflect.get(target, property);
         },
 
         set(target: Agent, property: PropertyKey, value: any) {
