+++ test/mocha.opts
@@ -1,4 +1,4 @@
 --compilers ts-node/register
 --require source-map-support/register
 --full-trace
-test/**/*.test.ts
+test/demo.test.ts
